% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fairCARA.R
\name{fairCARA}
\alias{fairCARA}
\title{Fair CARA experiments simulation}
\usage{
fairCARA(m, n, envy_c)
}
\arguments{
\item{m}{Estimates subgroup treatment effects}

\item{n}{Group-level conditional variance under treatment arm}

\item{envy_c}{Envyfreeness parameter}
}
\value{
\item{tau_opt}{Estimated treatment effects under proposed design}
\item{tau_cr}{Estimated treatment effects under complete randomization}
\item{tau_db}{Estimated treatment effects under doubly-adaptive biased coin design}
\item{sub_alloc_opt}{Treatment allocation under proposed design}
\item{sub_alloc_cr}{Treatment allocation under complete randomization}
\item{sub_alloc_db}{Treatment allocation under doubly-adaptive biased coin design}
\item{ate_opt}{Estimated treatment effect under the proposed design}
\item{ate_cr}{Estimated treatment effect under complete randomization}
\item{ate_db}{Estimated treatment effect under doubly adaptive biased coin design}
\item{cover_opt}{Coverage under the proposed design}
\item{cover_cr}{Coverage under complete randomization}
\item{cover_db}{Coverage under doubly adaptive biased coin design}
\item{power_opt}{Power under the proposed design}
\item{power_cr}{Power under complete randomization}
\item{power_db}{Power under doubly adaptive biased coin design}
}
\description{
This function simulate fair CARA experiments and return treatment effect estimates
}
