% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenS.R
\name{GenS}
\alias{GenS}
\title{Generate subgroup indices}
\usage{
GenS(X)
}
\arguments{
\item{X}{Covariates matrix}
}
\value{
Subgroup membership matrix
}
\description{
This function takes covariates mateix as an argument and generates a matrix of subgroup indices
}
