% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenX.R
\name{GenX}
\alias{GenX}
\title{Generate covariates}
\usage{
GenX(n)
}
\arguments{
\item{n}{Sample size}
}
\value{
Covariates matrix
}
\description{
This function takes sample size argument and generates a matrix of covariates
}
