% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenY.R
\name{GenY}
\alias{GenY}
\title{Generate outcomes}
\usage{
GenY(n, X, Tr, S)
}
\arguments{
\item{n}{Covariates matrix}

\item{X}{Covariates matrix}

\item{Tr}{Covariates matrix}

\item{S}{Covariates matrix}
}
\value{
Outcome vector
}
\description{
This function takes sample size, covariates matrix, treatments, and subgroup indices as arguments and generates a vector of outcomes
}
