% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubAlloc.R
\name{SubAlloc}
\alias{SubAlloc}
\title{Compute optimal treatment assignment probabilities}
\usage{
SubAlloc(tau, sigma1_vec, sigma0_vec, p_vec, N_t, envy_c, lower_c, upper_c)
}
\arguments{
\item{tau}{Estimates subgroup treatment effects}

\item{sigma1_vec}{Group-level conditional variance under treatment arm}

\item{sigma0_vec}{Group-level conditional variance under control arm}

\item{p_vec}{Subgroup proportions}

\item{N_t}{Number of subjects enrolled up to Stage t}

\item{envy_c}{Envyfreeness parameter}

\item{lower_c}{Lower bound on treatment assignment probabilities}

\item{upper_c}{Upper bound on treatment assignment probabilities}
}
\value{
Allocation vector
}
\description{
This function takes sample estimates and optimization constraints to generate a vector of treatment assignment probabilities
}
